Conflicts:
	counterpartylib/test/fixtures/rawtransactions.db
	counterpartylib/test/fixtures/scenarios/parseblock_unittest_fixture.json
	counterpartylib/test/fixtures/scenarios/parseblock_unittest_fixture.log
	counterpartylib/test/fixtures/scenarios/parseblock_unittest_fixture.sql
	counterpartylib/test/fixtures/scenarios/unittest_fixture.json
	counterpartylib/test/fixtures/scenarios/unittest_fixture.log
	counterpartylib/test/fixtures/scenarios/unittest_fixture.sql
	counterpartylib/test/fixtures/vectors.py
	counterpartylib/test/parse_block_test.py

+++ counterpartylib/test/util_test.py
@@ -19,10 +19,7 @@ import apsw
 import pytest
 import binascii
 import appdirs
-import pycoin
-from pycoin.tx import Tx
 import bitcoin as bitcoinlib
-logger = logging.getLogger(__name__)
 
 CURR_DIR = os.path.dirname(os.path.realpath(os.path.join(os.getcwd(), os.path.expanduser(__file__))))
 sys.path.append(os.path.normpath(os.path.join(CURR_DIR, '..')))
@@ -51,9 +48,6 @@ COUNTERPARTYD_OPTIONS = {
     'p2sh_dust_return_pubkey': '11' * 33,
     'utxo_locks_max_addresses': 0  # Disable UTXO locking for base test suite runs
 }
-
-# used for mocking the UTXO in various places, is automatically reset every test case
-MOCK_UTXO_SET = None
 
 
 def init_database(sqlfile, dbfile, options=None):
@@ -125,7 +119,6 @@ def remove_database_files(database_filename):
         if os.path.isfile(path):
             os.remove(path)
 
-
 def insert_block(db, block_index, parse_block=True):
     """Add blocks to the blockchain."""
     cursor = db.cursor()
@@ -139,15 +132,10 @@ def insert_block(db, block_index, parse_block=True):
 
     if parse_block:
         blocks.parse_block(db, block_index, block_time)
-
-    MOCK_UTXO_SET.increment_confirmations()
-
     return block_index, block_hash, block_time
-
 
 def create_next_block(db, block_index=None, parse_block=True):
     """Create faux data for the next block."""
-
     cursor = db.cursor()
     last_block_index = list(cursor.execute("SELECT block_index FROM blocks ORDER BY block_index DESC LIMIT 1"))[0]['block_index']
     if not block_index:
@@ -158,20 +146,22 @@ def create_next_block(db, block_index=None, parse_block=True):
 
     for index in range(last_block_index + 1, block_index + 1):
         inserted_block_index, block_hash, block_time = insert_block(db, index, parse_block=parse_block)
-
     cursor.close()
     return inserted_block_index, block_hash, block_time
 
-
-def insert_raw_transaction(raw_transaction, db):
+def insert_raw_transaction(raw_transaction, db, rawtransactions_db):
     """Add a raw transaction to the database."""
-    cursor = db.cursor()
-
     # one transaction per block
     block_index, block_hash, block_time = create_next_block(db, parse_block=False)
 
-    tx_hash = dummy_tx_hash(raw_transaction)
+    cursor = db.cursor()
     tx_index = block_index - config.BURN_START + 1
+
+    tx_hash = hashlib.sha256('{}{}'.format(tx_index,raw_transaction).encode('utf-8')).hexdigest()
+    # print(tx_hash)
+    # Remember to add it to the log dump
+    if pytest.config.option.savescenarios:
+        save_rawtransaction(rawtransactions_db, tx_hash, raw_transaction)
 
     source, destination, btc_amount, fee, data = blocks._get_tx_info(raw_transaction)
     transaction = (tx_index, tx_hash, block_index, block_hash, block_time, source, destination, btc_amount, fee, data, True)
@@ -179,79 +169,12 @@ def insert_raw_transaction(raw_transaction, db):
     tx = list(cursor.execute('''SELECT * FROM transactions WHERE tx_index = ?''', (tx_index,)))[0]
     cursor.close()
 
-    MOCK_UTXO_SET.add_raw_transaction(raw_transaction, tx_id=tx_hash, confirmations=1)
-
-    util.CURRENT_BLOCK_INDEX = block_index
+    util.CURRENT_BLOCK_INDEX = block_index  # TODO: Correct?!
     blocks.parse_block(db, block_index, block_time)
     return tx
 
-
-def insert_unconfirmed_raw_transaction(raw_transaction, db):
-    """Add a raw transaction to the database."""
-    # one transaction per block
-    cursor = db.cursor()
-
-    tx_hash = dummy_tx_hash(raw_transaction)
-
-    # this isn't really correct, but it will do
-    tx_index = list(cursor.execute('''SELECT tx_index FROM transactions ORDER BY tx_index DESC LIMIT 1'''))
-    tx_index = tx_index[0]['tx_index'] if len(tx_index) else 0
-    tx_index = tx_index + 1
-
-    source, destination, btc_amount, fee, data = blocks._get_tx_info(raw_transaction)
-    tx = {
-        'tx_index': tx_index,
-        'tx_hash': tx_hash,
-        'block_index': config.MEMPOOL_BLOCK_INDEX,
-        'block_hash': config.MEMPOOL_BLOCK_HASH,
-        'block_time': int(time.time()),
-        'source': source,
-        'destination': destination,
-        'btc_amount': btc_amount,
-        'fee': fee,
-        'data': data,
-        'supported': True
-    }
-
-    cursor.close()
-
-    MOCK_UTXO_SET.add_raw_transaction(raw_transaction, tx_id=tx_hash, confirmations=0)
-
-    return tx
-
-
-UNIQUE_DUMMY_TX_HASH = {}
-def dummy_tx_hash(raw_transaction):
-    global UNIQUE_DUMMY_TX_HASH
-
-    tx = pycoin.tx.Tx.from_hex(raw_transaction)
-
-    # normalize inputs
-    for txin in tx.txs_in:
-        txin.previous_hash = b'\x00' * 32
-        txin.previous_index = 0
-        txin.script = b'\x00'
-
-    # check if we have a change output
-    if tx.txs_out[-1].coin_value not in [0, config.DEFAULT_REGULAR_DUST_SIZE, config.DEFAULT_MULTISIG_DUST_SIZE, config.DEFAULT_OP_RETURN_VALUE]:
-        tx.txs_out[-1].coin_value = 0  # set change to 0
-
-    tx_id = tx.id()
-
-    # check we haven't created this before (if we do 2 exactly the sends for example)
-    if tx_id in UNIQUE_DUMMY_TX_HASH:
-        logger.warn('BUMP TXID %s' % tx_id)
-        UNIQUE_DUMMY_TX_HASH[tx_id] += 1
-        tx_id = hashlib.sha256('{}{}'.format(tx_id, UNIQUE_DUMMY_TX_HASH[tx_id]).encode('utf-8')).hexdigest()
-    else:
-        UNIQUE_DUMMY_TX_HASH[tx_id] = 0
-
-    return tx_id
-
-
 def insert_transaction(transaction, db):
     """Add a transaction to the database."""
-
     cursor = db.cursor()
     block = (transaction['block_index'], transaction['block_hash'], transaction['block_time'], None, None, None, None)
     cursor.execute('''INSERT INTO blocks (block_index, block_hash, block_time, ledger_hash, txlist_hash, previous_block_hash, difficulty) 
@@ -263,26 +186,19 @@ def insert_transaction(transaction, db):
 
 
 def initialise_rawtransactions_db(db):
-    cursor = db.cursor()
-    cursor.execute('CREATE TABLE IF NOT EXISTS raw_transactions(tx_hash TEXT UNIQUE, tx_hex TEXT)')
-    cursor.close()
-
-    prefill_rawtransactions_db(db)
-
-
-def prefill_rawtransactions_db(db):
     """Drop old raw transaction table, create new one and populate it from unspent_outputs.json."""
-    cursor = db.cursor()
-    cursor.execute('DROP TABLE  IF EXISTS raw_transactions')
-    cursor.execute('CREATE TABLE IF NOT EXISTS raw_transactions(tx_hash TEXT UNIQUE, tx_hex TEXT)')
-    with open(CURR_DIR + '/fixtures/unspent_outputs.json', 'r') as listunspent_test_file:
-            wallet_unspent = json.load(listunspent_test_file)
-            for output in wallet_unspent:
-                txid = binascii.hexlify(bitcoinlib.core.lx(output['txid'])).decode()
-                tx = backend.deserialize(output['txhex'])
-                cursor.execute('INSERT INTO raw_transactions VALUES (?, ?)', (txid, output['txhex']))
-    cursor.close()
-
+    if pytest.config.option.savescenarios:
+        server.initialise(database_file=':memory:', testnet=True, **COUNTERPARTYD_OPTIONS)
+        cursor = db.cursor()
+        cursor.execute('DROP TABLE  IF EXISTS raw_transactions')
+        cursor.execute('CREATE TABLE IF NOT EXISTS raw_transactions(tx_hash TEXT UNIQUE, tx_hex TEXT)')
+        with open(CURR_DIR + '/fixtures/unspent_outputs.json', 'r') as listunspent_test_file:
+                wallet_unspent = json.load(listunspent_test_file)
+                for output in wallet_unspent:
+                    txid = binascii.hexlify(bitcoinlib.core.lx(output['txid'])).decode()
+                    tx = backend.deserialize(output['txhex'])
+                    cursor.execute('INSERT INTO raw_transactions VALUES (?, ?)', (txid, output['txhex']))
+        cursor.close()
 
 def save_rawtransaction(db, tx_hash, tx_hex):
     """Insert the raw transaction into the db."""
@@ -307,7 +223,7 @@ def initialise_db(db):
     blocks.initialise(db)
     insert_block(db, config.BLOCK_FIRST - 1, parse_block=True)
 
-def run_scenario(scenario):
+def run_scenario(scenario, rawtransactions_db):
     """Execute a scenario for integration test, returns a dump of the db, a json with raw transactions and the full log."""
     server.initialise(database_file=':memory:', testnet=True, **COUNTERPARTYD_OPTIONS)
     config.PREFIX = b'TESTXXXX'
@@ -338,7 +254,7 @@ def run_scenario(scenario):
             compose = getattr(module, 'compose')
             unsigned_tx_hex = transaction.construct(db, compose(db, *tx[1]), **tx[2])
             raw_transactions.append({tx[0]: unsigned_tx_hex})
-            insert_raw_transaction(unsigned_tx_hex, db)
+            insert_raw_transaction(unsigned_tx_hex, db, rawtransactions_db)
         else:
             create_next_block(db, block_index=config.BURN_START + tx[1], parse_block=tx[2] if len(tx) == 3 else True)
 
@@ -349,13 +265,9 @@ def run_scenario(scenario):
     check.CHECKPOINTS_TESTNET = checkpoints
     return dump, log, json.dumps(raw_transactions, indent=4)
 
-def save_scenario(scenario_name):
+def save_scenario(scenario_name, rawtransactions_db):
     """Save currently run scenario's output for comparison with the expected outputs."""
-    dump, log, raw_transactions = run_scenario(INTEGRATION_SCENARIOS[scenario_name][0])
-
-    save_scenario_output(scenario_name, dump, log, raw_transactions)
-
-def save_scenario_output(scenario_name, dump, log, raw_transactions):
+    dump, log, raw_transactions = run_scenario(INTEGRATION_SCENARIOS[scenario_name][0], rawtransactions_db)
     with open(CURR_DIR + '/fixtures/scenarios/' + scenario_name + '.new.sql', 'w') as f:
         f.writelines(dump)
     with open(CURR_DIR + '/fixtures/scenarios/' + scenario_name + '.new.log', 'w') as f:
@@ -592,3 +504,45 @@ def reparse(testnet=True):
                 old_txlist = get_block_txlist(prod_db, block['block_index'])
                 compare_strings(old_txlist, new_txlist)
             raise(e)
+
+
+class ConfigContext(object):
+    def __init__(self, **kwargs):
+        self.config = config
+        self._extend = kwargs
+        self._before = {}
+
+    def __enter__(self):
+        self._before = {}
+        for k in self._extend:
+            self._before[k] = vars(self.config)[k]
+            vars(self.config)[k] = self._extend[k]
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        for k in self._before:
+            vars(self.config)[k] = self._before[k]
+
+
+class MockProtocolChangesContext(object):
+    def __init__(self, **kwargs):
+        from counterpartylib.test.conftest import MOCK_PROTOCOL_CHANGES
+        self.mock_protocol_changes = MOCK_PROTOCOL_CHANGES
+        self._extend = kwargs
+        self._before = {}
+        self._before_empty = []
+
+    def __enter__(self):
+        self._before = {}
+        for k in self._extend:
+            if k in self.mock_protocol_changes:
+                self._before[k] = self.mock_protocol_changes[k]
+            else:
+                self._before_empty.append(k)
+
+            self.mock_protocol_changes[k] = self._extend[k]
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        for k in self._before:
+            self.mock_protocol_changes[k] = self._before[k]
+        for k in self._before_empty:
+            del self.mock_protocol_changes[k]
